Extract OperationPath
