+++ src/components/openapi/OperationPath.module.css
@@ -1,0 +1,4 @@
+.variableLinkParameter {
+  font-weight: bold;
+  color: var(--ifm-color-primary);
+}
