+++ src/components/openapi/OperationReference.module.css
@@ -1,8 +1,3 @@
-.variableLinkParameter {
-  font-weight: bold;
-  color: var(--ifm-color-primary);
-}
-
 .parameterList {
   padding: 0;
   list-style-type: none;
